# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Koperasi App API",
        "description": "RESTful API for Koperasi App mobile applications and integrations",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pengajuan-anggota": {
            "get": {
                "tags": [
                    "Pengajuan Anggota"
                ],
                "summary": "List member loan applications",
                "description": "Display a listing of the resource.",
                "operationId": "105318d5aafe1e026232caa911f5a77f",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Filter by start date (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Filter by end date (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of loan applications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Pinjaman"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Anggota not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pengajuan Anggota"
                ],
                "summary": "Create a new loan application",
                "description": "Store a newly created resource in storage.",
                "operationId": "9bcf54cd9cfba6ad2cf9badb34e2d9cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nominal",
                                    "tenor_bulan",
                                    "keperluan"
                                ],
                                "properties": {
                                    "nominal": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000000,
                                        "minimum": 1000000
                                    },
                                    "tenor_bulan": {
                                        "type": "integer",
                                        "example": 12,
                                        "maximum": 120,
                                        "minimum": 1
                                    },
                                    "keperluan": {
                                        "type": "string",
                                        "example": "Renovasi rumah",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Loan application created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pinjaman"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Only active members can submit applications"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pengajuan-anggota/{id}": {
            "get": {
                "tags": [
                    "Pengajuan Anggota"
                ],
                "summary": "Get loan application details",
                "description": "Display the specified resource.",
                "operationId": "036bf00f72f25d834c0a613e3b9cd882",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Loan application ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Loan application details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pinjaman"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Unauthorized - application does not belong to you"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pengajuan Anggota"
                ],
                "summary": "Update loan application (only if pending)",
                "description": "Update the specified resource in storage.",
                "operationId": "8bc500f9ee8a3c57edaa1a53a5ce9276",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Loan application ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nominal": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000000,
                                        "minimum": 1000000
                                    },
                                    "tenor_bulan": {
                                        "type": "integer",
                                        "example": 12,
                                        "maximum": 120,
                                        "minimum": 1
                                    },
                                    "keperluan": {
                                        "type": "string",
                                        "example": "Renovasi rumah",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Loan application updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pinjaman"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Can only update applications that are still pending"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pengajuan Anggota"
                ],
                "summary": "Delete loan application (only if pending)",
                "description": "Remove the specified resource from storage.",
                "operationId": "b4cc57d2ddc9eef536ca23a4b61b646a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Loan application ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Application deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Can only delete applications that are still pending"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "no_karyawan": {
                        "type": "string",
                        "example": "TOP-100001"
                    },
                    "sso_id": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "auth_provider": {
                        "type": "string",
                        "example": "local"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "harus_ganti_password": {
                        "type": "boolean",
                        "example": false
                    },
                    "status": {
                        "type": "string",
                        "example": "aktif"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Pinjaman": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "anggota_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "pengaju_user_id": {
                        "type": "integer",
                        "example": 4
                    },
                    "nominal": {
                        "type": "string",
                        "example": "5000000.00"
                    },
                    "tenor_bulan": {
                        "type": "integer",
                        "example": 12
                    },
                    "keperluan": {
                        "type": "string",
                        "example": "Renovasi rumah"
                    },
                    "snapshot_bank": {
                        "type": "string",
                        "example": "BCA",
                        "nullable": true
                    },
                    "snapshot_no_rekening": {
                        "type": "string",
                        "example": "1234567890",
                        "nullable": true
                    },
                    "snapshot_atas_nama": {
                        "type": "string",
                        "example": "John Doe",
                        "nullable": true
                    },
                    "persentase_bunga": {
                        "type": "string",
                        "example": "1.00"
                    },
                    "status": {
                        "type": "string",
                        "example": "diajukan",
                        "enum": [
                            "diajukan",
                            "approved_bendahara",
                            "approved_ketua",
                            "aktif",
                            "ditolak",
                            "lunas"
                        ]
                    },
                    "cair_oleh_bendahara": {
                        "type": "boolean",
                        "example": false
                    },
                    "sudah_pakai_privilege_reloan": {
                        "type": "boolean",
                        "example": false
                    },
                    "sudah_pakai_percepatan": {
                        "type": "boolean",
                        "example": false
                    },
                    "tanggal_pengajuan": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "tanggal_pencairan": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "disetujui_pada": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "versi_syarat": {
                        "type": "string",
                        "nullable": true
                    },
                    "ip_address_setuju": {
                        "type": "string",
                        "nullable": true
                    },
                    "user_agent_setuju": {
                        "type": "string",
                        "nullable": true
                    },
                    "catatan_bendahara": {
                        "type": "string",
                        "nullable": true
                    },
                    "catatan_ketua": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "anggota": {
                        "$ref": "#/components/schemas/Anggota"
                    },
                    "pengaju": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Anggota": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 4
                    },
                    "no_anggota": {
                        "type": "string",
                        "example": "ANG-2026-0001"
                    },
                    "no_karyawan": {
                        "type": "string",
                        "example": "TOP-100001"
                    },
                    "no_ktp": {
                        "type": "string",
                        "nullable": true
                    },
                    "nama": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "cabang": {
                        "type": "string",
                        "example": "Banjarmasin"
                    },
                    "unit_bisnis": {
                        "type": "string",
                        "example": "Operasional"
                    },
                    "department": {
                        "type": "string",
                        "example": "Operasional"
                    },
                    "divisi": {
                        "type": "string",
                        "example": "Lapangan"
                    },
                    "no_hp": {
                        "type": "string",
                        "nullable": true
                    },
                    "alamat": {
                        "type": "string",
                        "nullable": true
                    },
                    "jabatan": {
                        "type": "string",
                        "example": "staff"
                    },
                    "tanggal_mulai_kerja": {
                        "type": "string",
                        "format": "date"
                    },
                    "tanggal_jadi_anggota": {
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "type": "string",
                        "example": "aktif"
                    },
                    "tanggal_resign": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "alasan_resign": {
                        "type": "string",
                        "nullable": true
                    },
                    "resigned_by": {
                        "type": "integer",
                        "nullable": true
                    },
                    "resigned_settlement_json": {
                        "type": "array",
                        "items": {},
                        "nullable": true
                    },
                    "reaktivasi_history_json": {
                        "type": "array",
                        "items": {},
                        "nullable": true
                    },
                    "limit_custom": {
                        "type": "string",
                        "nullable": true
                    },
                    "limit_custom_keterangan": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}